add missing snapshots
