+++ packages/tscc/test/transformer/transformers.ts
@@ -1,9 +1,10 @@
 ///<reference types="jest" />
 import * as ts from 'typescript';
+import * as tsickle from 'tsickle';
+import TsccSpecWithTS from '../../src/spec/TsccSpecWithTS';
 import decoratorPropertyTransformer from '../../src/transformer/decoratorPropertyTransformer';
 import restPropertyTransformer from '../../src/transformer/restPropertyTransformer';
 import path = require('path');
-import {emit} from './test_harness';
 
 const samplesRoot = path.resolve(__dirname, '../sample');
 
@@ -103,3 +104,32 @@ describe(`restPropertyTransformer`, function () {
 		})
 	})
 });
+
+type EmitTransformerFactory = (host: tsickle.TsickleHost) => tsickle.EmitTransformers;
+
+function emit(tsconfigPath: string, files: string[], transformerFactory: EmitTransformerFactory, override: Partial<ts.CompilerOptions> = {}) {
+	const {parsedConfig} = TsccSpecWithTS.loadTsConfigFromPath(tsconfigPath);
+	const {options} = parsedConfig;
+	Object.assign(options, override);
+	const host = ts.createCompilerHost(options);
+	const program = ts.createProgram(files, options, host);
+	const transformerHost = {
+		shouldSkipTsickleProcessing: () => false,
+		shouldIgnoreWarningsForPath: () => true,
+		googmodule: false,
+		pathToModuleName: x => x,
+		fileNameToModuleId: x => x,
+		moduleResolutionHost: ts.createCompilerHost(parsedConfig.options),
+		options: parsedConfig.options,
+	};
+
+	const out = new Map<string, string>();
+	const {externs} = tsickle.emitWithTsickle(
+		program, transformerHost, host, options, undefined,
+		(fileName, data) => {out.set(fileName, data);},
+		undefined /* cancellationtoken */, false /* emitOnlyDtsFiles */,
+		transformerFactory(transformerHost)
+	);
+	return {out, externs}
+}
+
